+++ services/web/server/src/simcore_service_webserver/meta_modeling/_projects.py
@@ -17,11 +17,8 @@ from models_library.projects import ProjectID
 from .._meta import api_version_prefix as VTAG
 from ..director_v2_api import AbstractProjectRunPolicy
 from ..projects.projects_handlers_crud import RQ_REQUESTED_REPO_PROJECT_UUID_KEY
+from ._iterations import get_or_create_runnable_projects, get_runnable_projects_ids
 from ._version_control import CommitID, VersionControlForMetaModeling
-from .meta_modeling_iterations import (
-    get_or_create_runnable_projects,
-    get_runnable_projects_ids,
-)
 
 log = logging.getLogger(__name__)
