+++ services/web/server/tests/unit/with_dbs/03/meta_modeling/test_meta_modeling_handlers.py
@@ -6,12 +6,12 @@
 import pytest
 from openapi_core.schema.specs.models import Spec as OpenApiSpecs
 from simcore_service_webserver._meta import API_VTAG as VX
-from simcore_service_webserver.meta_modeling import meta_modeling_handlers
+from simcore_service_webserver.meta_modeling import _rest_handlers
 
 
 @pytest.mark.parametrize(
     "route",
-    meta_modeling_handlers.routes,
+    _rest_handlers.routes,
     ids=lambda r: f"{r.method.upper()} {r.path}",
 )
 def test_route_against_openapi_specs(route, openapi_specs: OpenApiSpecs):
